Split PSF kernels from CFW logic
